+++ src/main/java/com/woowacamp/soolsool/core/liquor/controller/LiquorStockController.java
@@ -1,6 +1,6 @@
 package com.woowacamp.soolsool.core.liquor.controller;
 
-import com.woowacamp.soolsool.core.liquor.dto.LiquorStockSaveRequest;
+import com.woowacamp.soolsool.core.liquor.dto.request.LiquorStockSaveRequest;
 import com.woowacamp.soolsool.core.liquor.service.LiquorStockService;
 import com.woowacamp.soolsool.global.aop.RequestLogging;
 import com.woowacamp.soolsool.global.auth.dto.Vendor;
