+++ src/main/java/com/woowacamp/soolsool/core/liquor/dto/request/LiquorStockSaveRequest.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.liquor.dto;
+package com.woowacamp.soolsool.core.liquor.dto.request;
 
 import com.woowacamp.soolsool.core.liquor.domain.LiquorStock;
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorStockCount;
