+++ src/main/java/com/woowacamp/soolsool/core/liquor/dto/response/LiquorElementResponse.java
@@ -1,8 +1,7 @@
-package com.woowacamp.soolsool.core.liquor.dto;
+package com.woowacamp.soolsool.core.liquor.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.woowacamp.soolsool.core.liquor.domain.Liquor;
-import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorCtrClick;
 import lombok.Getter;
 
 @Getter
@@ -13,27 +12,24 @@ public class LiquorElementResponse {
     private final String price;
     private final String imageUrl;
     private final Integer stock;
-    private final Long clickCount;
 
     @JsonCreator
     public LiquorElementResponse(final Long id, final String name, final String price,
-        final String imageUrl, final Integer stock, final Long clickCount
+        final String imageUrl, final Integer stock
     ) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
         this.stock = stock;
-        this.clickCount = clickCount;
     }
 
-    public LiquorElementResponse(final Liquor liquor, final LiquorCtrClick clickCount) {
+    public LiquorElementResponse(final Liquor liquor) {
         this.id = liquor.getId();
         this.name = liquor.getName();
         this.price = liquor.getPrice().toString();
         this.imageUrl = liquor.getImageUrl();
         this.stock = liquor.getTotalStock();
-        this.clickCount = clickCount.getCount();
     }
 
     public static LiquorElementResponse from(final Liquor liquor) {
@@ -42,18 +38,16 @@ public class LiquorElementResponse {
             liquor.getName(),
             liquor.getPrice().toString(),
             liquor.getImageUrl(),
-            liquor.getTotalStock(),
-            0L
+            liquor.getTotalStock()
         );
     }
-    public static LiquorElementResponse of(final Liquor liquor, final Long clickCount) {
+    public static LiquorElementResponse of(final Liquor liquor) {
         return new LiquorElementResponse(
             liquor.getId(),
             liquor.getName(),
             liquor.getPrice().toString(),
             liquor.getImageUrl(),
-            liquor.getTotalStock(),
-            clickCount
+            liquor.getTotalStock()
         );
     }
 }
