+++ src/main/java/com/woowacamp/soolsool/core/liquor/repository/LiquorQueryDslRepository.java
@@ -10,8 +10,8 @@ import com.woowacamp.soolsool.core.liquor.domain.LiquorBrew;
 import com.woowacamp.soolsool.core.liquor.domain.LiquorRegion;
 import com.woowacamp.soolsool.core.liquor.domain.LiquorStatus;
 import com.woowacamp.soolsool.core.liquor.domain.vo.LiquorBrand;
-import com.woowacamp.soolsool.core.liquor.dto.LiquorElementResponse;
-import com.woowacamp.soolsool.core.liquor.dto.LiquorSearchCondition;
+import com.woowacamp.soolsool.core.liquor.dto.response.LiquorElementResponse;
+import com.woowacamp.soolsool.core.liquor.dto.request.LiquorSearchCondition;
 import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
