+++ src/test/java/com/woowacamp/soolsool/core/liquor/service/LiquorStockServiceIntegrationTest.java
@@ -2,7 +2,7 @@ package com.woowacamp.soolsool.core.liquor.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.woowacamp.soolsool.core.liquor.dto.LiquorStockSaveRequest;
+import com.woowacamp.soolsool.core.liquor.dto.request.LiquorStockSaveRequest;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
